feat: move assets methods to the parent class
